added save and load library feature, changed all float to double type

+++ certoscale/src/main/java/com/certoclav/certoscale/listener/WeightListener.java
@@ -2,5 +2,5 @@ package com.certoclav.certoscale.listener;
 
 
 public interface WeightListener {
- void onWeightChanged(Float weight, String unit);
+ void onWeightChanged(Double weight, String unit);
 }
